Start nukkit sign plugin
