+++ cloudnet-modules/cloudnet-signs/src/main/java/de/dytanic/cloudnet/ext/signs/AbstractSignManagement.java
@@ -1,44 +1,391 @@
 package de.dytanic.cloudnet.ext.signs;
 
 import de.dytanic.cloudnet.common.Validate;
+import de.dytanic.cloudnet.common.collection.Iterables;
+import de.dytanic.cloudnet.common.collection.Maps;
+import de.dytanic.cloudnet.common.collection.Pair;
 import de.dytanic.cloudnet.common.concurrent.ITask;
 import de.dytanic.cloudnet.common.document.gson.JsonDocument;
+import de.dytanic.cloudnet.common.unsafe.CPUUsageResolver;
 import de.dytanic.cloudnet.driver.CloudNetDriver;
 import de.dytanic.cloudnet.driver.network.def.PacketConstants;
 import de.dytanic.cloudnet.driver.service.ServiceEnvironmentType;
 import de.dytanic.cloudnet.driver.service.ServiceInfoSnapshot;
+import de.dytanic.cloudnet.driver.service.ServiceLifeCycle;
+import de.dytanic.cloudnet.driver.service.ServiceTemplate;
+import de.dytanic.cloudnet.wrapper.Wrapper;
 
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Function;
 
 public abstract class AbstractSignManagement {
 
-    protected static AbstractSignManagement instance;
+    private static final Comparator<Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>>>
+            ENTRY_COMPARATOR = new ServiceInfoSnapshotEntryComparator(),
+            ENTRY_COMPARATOR_2 = new ServiceInfoSnapshotEntryComparator2();
 
+    private static AbstractSignManagement instance;
+    private final AtomicInteger[] indexes = new AtomicInteger[]{
+            new AtomicInteger(-1), //starting
+            new AtomicInteger(-1) //search
+    };
+    private final Map<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>> services = Maps.newConcurrentHashMap();
     protected List<Sign> signs;
 
-    public static AbstractSignManagement getInstance() {
-        return AbstractSignManagement.instance;
+    public AbstractSignManagement() {
+        instance = this;
+
+        this.signs = Iterables.newCopyOnWriteArrayList();
+        this.signs.addAll(this.getSignsFromNode());
+
+        this.executeSearchingTask();
+        this.executeStartingTask();
     }
 
-    public abstract void onRegisterService(ServiceInfoSnapshot serviceInfoSnapshot);
+    public static AbstractSignManagement getInstance() {
+        return instance;
+    }
 
-    public abstract void onStartService(ServiceInfoSnapshot serviceInfoSnapshot);
+    protected abstract void updateSignNext(Sign sign, SignLayout signLayout, ServiceInfoSnapshot serviceInfoSnapshot);
 
-    public abstract void onConnectService(ServiceInfoSnapshot serviceInfoSnapshot);
+    protected abstract void runTaskLater(Runnable runnable, long delay);
 
-    public abstract void onUpdateServiceInfo(ServiceInfoSnapshot serviceInfoSnapshot);
+    private void putService(ServiceInfoSnapshot serviceInfoSnapshot, Function<SignConfigurationEntry, ServiceInfoState> stateFunction) {
+        if (!this.isImportantCloudService(serviceInfoSnapshot)) {
+            return;
+        }
 
-    public abstract void onDisconnectService(ServiceInfoSnapshot serviceInfoSnapshot);
+        SignConfigurationEntry entry = this.getOwnSignConfigurationEntry();
+        if (entry == null) {
+            return;
+        }
 
-    public abstract void onStopService(ServiceInfoSnapshot serviceInfoSnapshot);
+        this.services.put(serviceInfoSnapshot.getServiceId().getUniqueId(), new Pair<>(serviceInfoSnapshot, stateFunction.apply(entry)));
+        this.updateSigns();
+    }
 
-    public abstract void onUnregisterService(ServiceInfoSnapshot serviceInfoSnapshot);
+    public void onRegisterService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        this.putService(serviceInfoSnapshot, entry -> ServiceInfoState.STOPPED);
+    }
 
-    public abstract void onSignAdd(Sign sign);
+    public void onStartService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        this.putService(serviceInfoSnapshot, entry -> ServiceInfoState.STARTING);
+    }
 
-    public abstract void onSignRemove(Sign sign);
+    public void onConnectService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        this.putService(serviceInfoSnapshot, entry -> this.fromServiceInfoSnapshot(serviceInfoSnapshot, entry));
+    }
+
+    public void onUpdateServiceInfo(ServiceInfoSnapshot serviceInfoSnapshot) {
+        this.putService(serviceInfoSnapshot, entry -> this.fromServiceInfoSnapshot(serviceInfoSnapshot, entry));
+    }
+
+    public void onDisconnectService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        this.putService(serviceInfoSnapshot, entry -> ServiceInfoState.STOPPED);
+    }
+
+    public void onStopService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        this.putService(serviceInfoSnapshot, entry -> ServiceInfoState.STOPPED);
+    }
+
+    public void onUnregisterService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        if (!this.isImportantCloudService(serviceInfoSnapshot)) {
+            return;
+        }
+
+        SignConfigurationEntry entry = this.getOwnSignConfigurationEntry();
+        if (entry == null) {
+            return;
+        }
+
+        this.services.remove(serviceInfoSnapshot.getServiceId().getUniqueId());
+        this.updateSigns();
+    }
+
+    public void onSignAdd(Sign sign) {
+        Validate.checkNotNull(sign);
+
+        this.signs.add(sign);
+        CloudNetDriver.getInstance().getTaskScheduler().schedule(this::updateSigns);
+    }
+
+    public void onSignRemove(Sign sign) {
+        Validate.checkNotNull(sign);
+
+        Sign signEntry = Iterables.first(this.signs, s -> s.getSignId() == sign.getSignId());
+
+        if (signEntry != null) {
+            this.signs.remove(signEntry);
+        }
+
+        CloudNetDriver.getInstance().getTaskScheduler().schedule(this::updateSigns);
+    }
+
+    public void updateSigns() {
+        SignConfigurationEntry signConfiguration = this.getOwnSignConfigurationEntry();
+        if (signConfiguration == null) {
+            return;
+        }
+
+        List<Sign> signs = Iterables.newArrayList(Iterables.filter(this.signs, sign -> Iterables.contains(sign.getProvidedGroup(), Wrapper.getInstance().getServiceConfiguration().getGroups())));
+
+        Collections.sort(signs);
+        if (signs.isEmpty()) {
+            return;
+        }
+
+        List<Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>>> cachedFilter = Iterables.newArrayList();
+        List<Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>>> entries = Iterables.newArrayList(Iterables.filter(this.services.entrySet(),
+                item -> item.getValue().getSecond() != ServiceInfoState.STOPPED));
+
+        entries.sort(ENTRY_COMPARATOR);
+
+        for (Sign sign : signs) {
+            this.updateSign(sign, signConfiguration, cachedFilter, entries);
+        }
+    }
+
+    private void updateSign(Sign sign,
+                            SignConfigurationEntry signConfiguration,
+                            List<Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>>> cachedFilter,
+                            List<Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>>> entries) {
+        Iterables.filter(entries, entry -> {
+            boolean access = Iterables.contains(sign.getTargetGroup(), entry.getValue().getFirst().getConfiguration().getGroups());
+
+            if (sign.getTemplatePath() != null) {
+                boolean condition = false;
+
+                for (ServiceTemplate template : entry.getValue().getFirst().getConfiguration().getTemplates()) {
+                    if (sign.getTemplatePath().equals(template.getTemplatePath())) {
+                        condition = true;
+                        break;
+                    }
+                }
+
+                access = condition;
+            }
+
+            return access;
+        }, cachedFilter);
+
+        cachedFilter.sort(ENTRY_COMPARATOR_2);
+
+        if (!cachedFilter.isEmpty()) {
+            Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>> entry = cachedFilter.get(0);
+
+            sign.setServiceInfoSnapshot(entry.getValue().getFirst());
+
+            this.applyState(sign, signConfiguration, entry.getValue().getFirst(), entry.getValue().getSecond());
+
+            entries.remove(entry);
+
+        } else {
+            sign.setServiceInfoSnapshot(null);
+
+            if (!signConfiguration.getSearchLayouts().getSignLayouts().isEmpty()) {
+                this.updateSignNext(sign, signConfiguration.getSearchLayouts().getSignLayouts().get(this.indexes[1].get()), null);
+            }
+        }
+
+        cachedFilter.clear();
+    }
+
+    private void applyState(Sign sign, SignConfigurationEntry signConfiguration, ServiceInfoSnapshot serviceInfoSnapshot, ServiceInfoState state) {
+        switch (state) {
+            case STOPPED: {
+                sign.setServiceInfoSnapshot(null);
+
+                if (!signConfiguration.getSearchLayouts().getSignLayouts().isEmpty()) {
+                    this.updateSignNext(sign, signConfiguration.getSearchLayouts().getSignLayouts().get(this.indexes[1].get()), null);
+                }
+            }
+            break;
+            case STARTING: {
+                sign.setServiceInfoSnapshot(null);
+
+                if (!signConfiguration.getStartingLayouts().getSignLayouts().isEmpty()) {
+                    this.updateSignNext(sign, signConfiguration.getStartingLayouts().getSignLayouts().get(this.indexes[0].get()), serviceInfoSnapshot);
+                }
+            }
+            break;
+            case EMPTY_ONLINE: {
+                SignLayout signLayout = null;
+
+                SignConfigurationTaskEntry taskEntry = this.getValidSignConfigurationTaskEntryFromSignConfigurationEntry(signConfiguration, sign.getTargetGroup());
+
+                if (taskEntry != null) {
+                    signLayout = taskEntry.getEmptyLayout();
+                }
+
+                if (signLayout == null) {
+                    signLayout = signConfiguration.getDefaultEmptyLayout();
+                }
+
+                this.updateSignNext(sign, signLayout, serviceInfoSnapshot);
+            }
+            break;
+            case ONLINE: {
+                SignLayout signLayout = null;
+
+                SignConfigurationTaskEntry taskEntry = this.getValidSignConfigurationTaskEntryFromSignConfigurationEntry(signConfiguration, sign.getTargetGroup());
+
+                if (taskEntry != null) {
+                    signLayout = taskEntry.getOnlineLayout();
+                }
+
+                if (signLayout == null) {
+                    signLayout = signConfiguration.getDefaultOnlineLayout();
+                }
+
+                this.updateSignNext(sign, signLayout, serviceInfoSnapshot);
+            }
+            break;
+            case FULL_ONLINE: {
+                SignLayout signLayout = null;
+
+                SignConfigurationTaskEntry taskEntry = this.getValidSignConfigurationTaskEntryFromSignConfigurationEntry(signConfiguration, sign.getTargetGroup());
+
+                if (taskEntry != null) {
+                    signLayout = taskEntry.getFullLayout();
+                }
+
+                if (signLayout == null) {
+                    signLayout = signConfiguration.getDefaultFullLayout();
+                }
+
+                this.updateSignNext(sign, signLayout, serviceInfoSnapshot);
+            }
+            break;
+        }
+    }
+
+    protected String addDataToLine(Sign sign, String input, ServiceInfoSnapshot serviceInfoSnapshot) {
+        Validate.checkNotNull(input);
+
+        if (serviceInfoSnapshot == null) {
+            return input;
+        }
+
+        input = input.replace("%task%", serviceInfoSnapshot.getServiceId().getTaskName());
+        input = input.replace("%task_id%", String.valueOf(serviceInfoSnapshot.getServiceId().getTaskServiceId()));
+        input = input.replace("%group%", sign.getTargetGroup());
+        input = input.replace("%name%", serviceInfoSnapshot.getServiceId().getName());
+        input = input.replace("%uuid%", serviceInfoSnapshot.getServiceId().getUniqueId().toString().split("-")[0]);
+        input = input.replace("%node%", serviceInfoSnapshot.getServiceId().getNodeUniqueId());
+        input = input.replace("%environment%", String.valueOf(serviceInfoSnapshot.getServiceId().getEnvironment()));
+        input = input.replace("%life_cycle%", String.valueOf(serviceInfoSnapshot.getLifeCycle()));
+        input = input.replace("%runtime%", serviceInfoSnapshot.getConfiguration().getRuntime());
+        input = input.replace("%port%", String.valueOf(serviceInfoSnapshot.getConfiguration().getPort()));
+        input = input.replace("%cpu_usage%", CPUUsageResolver.CPU_USAGE_OUTPUT_FORMAT.format(serviceInfoSnapshot.getProcessSnapshot().getCpuUsage()));
+        input = input.replace("%threads%", String.valueOf(serviceInfoSnapshot.getProcessSnapshot().getThreads().size()));
+        input = input.replace("%online%",
+                (serviceInfoSnapshot.getProperties().contains("Online") && serviceInfoSnapshot.getProperties().getBoolean("Online")
+                        ? "Online" : "Offline"
+                ));
+        input = input.replace("%online_players%", String.valueOf(serviceInfoSnapshot.getProperties().getInt("Online-Count")));
+        input = input.replace("%max_players%", String.valueOf(serviceInfoSnapshot.getProperties().getInt("Max-Players")));
+        input = input.replace("%motd%", serviceInfoSnapshot.getProperties().getString("Motd"));
+        input = input.replace("%extra%", serviceInfoSnapshot.getProperties().getString("Extra"));
+        input = input.replace("%state%", serviceInfoSnapshot.getProperties().getString("State"));
+        input = input.replace("%version%", serviceInfoSnapshot.getProperties().getString("Version"));
+        input = input.replace("%whitelist%", (serviceInfoSnapshot.getProperties().contains("Whitelist-Enabled") &&
+                serviceInfoSnapshot.getProperties().getBoolean("Whitelist-Enabled")
+                ? "Enabled" : "Disabled"
+        ));
+
+        return input;
+    }
+
+    private SignConfigurationTaskEntry getValidSignConfigurationTaskEntryFromSignConfigurationEntry(SignConfigurationEntry entry, String targetTask) {
+        return Iterables.first(entry.getTaskLayouts(), signConfigurationTaskEntry -> signConfigurationTaskEntry.getTask() != null &&
+                signConfigurationTaskEntry.getEmptyLayout() != null &&
+                signConfigurationTaskEntry.getFullLayout() != null &&
+                signConfigurationTaskEntry.getOnlineLayout() != null &&
+                signConfigurationTaskEntry.getTask().equalsIgnoreCase(targetTask));
+    }
+
+    public SignConfigurationEntry getOwnSignConfigurationEntry() {
+        return Iterables.first(SignConfigurationProvider.load().getConfigurations(), signConfigurationEntry -> Iterables.contains(signConfigurationEntry.getTargetGroup(), Wrapper.getInstance().getServiceConfiguration().getGroups()));
+    }
+
+    private boolean isEmptyService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        return serviceInfoSnapshot.isConnected() &&
+                serviceInfoSnapshot.getProperties().getBoolean("Online") &&
+                serviceInfoSnapshot.getProperties().contains("Online-Count") &&
+                serviceInfoSnapshot.getProperties().getInt("Online-Count") == 0;
+    }
+
+    private boolean isFullService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        return serviceInfoSnapshot.isConnected() &&
+                serviceInfoSnapshot.getProperties().getBoolean("Online") &&
+                serviceInfoSnapshot.getProperties().contains("Online-Count") &&
+                serviceInfoSnapshot.getProperties().contains("Max-Players") &&
+                serviceInfoSnapshot.getProperties().getInt("Online-Count") >=
+                        serviceInfoSnapshot.getProperties().getInt("Max-Players");
+    }
+
+    private boolean isStartingService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        return serviceInfoSnapshot.getLifeCycle() == ServiceLifeCycle.RUNNING && !serviceInfoSnapshot.getProperties().contains("Online");
+    }
+
+    private boolean isIngameService(ServiceInfoSnapshot serviceInfoSnapshot) {
+        return serviceInfoSnapshot.getLifeCycle() == ServiceLifeCycle.RUNNING
+                && serviceInfoSnapshot.isConnected()
+                &&
+                serviceInfoSnapshot.getProperties().getBoolean("Online")
+                && (
+                (serviceInfoSnapshot.getProperties().contains("Motd") &&
+                        (
+                                serviceInfoSnapshot.getProperties().getString("Motd").toLowerCase().contains("ingame") ||
+                                        serviceInfoSnapshot.getProperties().getString("Motd").toLowerCase().contains("running")
+                        )
+                ) ||
+                        (serviceInfoSnapshot.getProperties().contains("Extra") &&
+                                (
+                                        serviceInfoSnapshot.getProperties().getString("Extra").toLowerCase().contains("ingame") ||
+                                                serviceInfoSnapshot.getProperties().getString("Extra").toLowerCase().contains("running")
+                                )
+                        ) ||
+                        (serviceInfoSnapshot.getProperties().contains("State") &&
+                                (
+                                        serviceInfoSnapshot.getProperties().getString("State").toLowerCase().contains("ingame") ||
+                                                serviceInfoSnapshot.getProperties().getString("State").toLowerCase().contains("running")
+                                )
+                        )
+        );
+    }
+
+    private ServiceInfoState fromServiceInfoSnapshot(ServiceInfoSnapshot serviceInfoSnapshot, SignConfigurationEntry signConfiguration) {
+        if (isIngameService(serviceInfoSnapshot)) {
+            return ServiceInfoState.STOPPED;
+        }
+
+        if (isEmptyService(serviceInfoSnapshot)) {
+            return ServiceInfoState.EMPTY_ONLINE;
+        }
+
+        if (this.isFullService(serviceInfoSnapshot)) {
+            if (!signConfiguration.isSwitchToSearchingWhenServiceIsFull()) {
+                return ServiceInfoState.FULL_ONLINE;
+            } else {
+                return ServiceInfoState.STOPPED;
+            }
+        }
+
+        if (this.isStartingService(serviceInfoSnapshot)) {
+            return ServiceInfoState.STARTING;
+        }
+
+        if (serviceInfoSnapshot.getLifeCycle() == ServiceLifeCycle.RUNNING &&
+                serviceInfoSnapshot.isConnected() &&
+                serviceInfoSnapshot.getProperties().getBoolean("Online")) {
+            return ServiceInfoState.ONLINE;
+        } else {
+            return ServiceInfoState.STOPPED;
+        }
+    }
 
     public void sendSignAddUpdate(Sign sign) {
         Validate.checkNotNull(sign);
@@ -62,7 +409,7 @@ public abstract class AbstractSignManagement {
                 );
     }
 
-    public Collection<Sign> getSignsFromNode() {
+    protected Collection<Sign> getSignsFromNode() {
         ITask<Collection<Sign>> signs = CloudNetDriver.getInstance().getPacketQueryProvider().sendCallablePacket(
                 CloudNetDriver.getInstance().getNetworkClient().getChannels().iterator().next(),
                 SignConstants.SIGN_CHANNEL_SYNC_CHANNEL_PROPERTY,
@@ -90,13 +437,75 @@ public abstract class AbstractSignManagement {
         );
     }
 
-    public boolean isImportantCloudService(ServiceInfoSnapshot serviceInfoSnapshot) {
+    private boolean isImportantCloudService(ServiceInfoSnapshot serviceInfoSnapshot) {
         return serviceInfoSnapshot != null &&
                 (
                         serviceInfoSnapshot.getServiceId().getEnvironment() == ServiceEnvironmentType.MINECRAFT_SERVER ||
                                 serviceInfoSnapshot.getServiceId().getEnvironment() == ServiceEnvironmentType.GLOWSTONE
                 )
                 ;
+    }
+
+    protected void executeStartingTask() {
+        SignConfigurationEntry signConfigurationEntry = getOwnSignConfigurationEntry();
+
+        if (signConfigurationEntry != null && signConfigurationEntry.getStartingLayouts() != null &&
+                signConfigurationEntry.getStartingLayouts().getSignLayouts().size() > 0) {
+            if (this.indexes[0].get() == -1) {
+                this.indexes[0].set(0);
+            }
+
+            if ((this.indexes[0].get() + 1) < signConfigurationEntry.getStartingLayouts().getSignLayouts().size()) {
+                this.indexes[0].incrementAndGet();
+            } else {
+                this.indexes[0].set(0);
+            }
+
+            this.runTaskLater(
+                    this::executeStartingTask,
+                    20 / (Math.min(signConfigurationEntry.getStartingLayouts().getAnimationsPerSecond(), 20))
+            );
+        } else {
+            this.indexes[0].set(-1);
+            this.runTaskLater(this::executeStartingTask, 20);
+        }
+
+        CloudNetDriver.getInstance().getTaskScheduler().schedule(this::updateSigns);
+    }
+
+    protected void executeSearchingTask() {
+        SignConfigurationEntry signConfigurationEntry = getOwnSignConfigurationEntry();
+
+        if (signConfigurationEntry != null && signConfigurationEntry.getSearchLayouts() != null &&
+                signConfigurationEntry.getSearchLayouts().getSignLayouts().size() > 0) {
+            if (this.indexes[1].get() == -1) {
+                this.indexes[1].set(0);
+            }
+
+            if ((this.indexes[1].get() + 1) < signConfigurationEntry.getSearchLayouts().getSignLayouts().size()) {
+                this.indexes[1].incrementAndGet();
+            } else {
+                this.indexes[1].set(0);
+            }
+
+            this.runTaskLater(
+                    this::executeSearchingTask,
+                    20 / (Math.min(signConfigurationEntry.getSearchLayouts().getAnimationsPerSecond(), 20))
+            );
+        } else {
+            this.indexes[1].set(-1);
+            this.runTaskLater(this::executeSearchingTask, 20);
+        }
+
+        CloudNetDriver.getInstance().getTaskScheduler().schedule(this::updateSigns);
+    }
+
+    public Map<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>> getServices() {
+        return this.services;
+    }
+
+    public AtomicInteger[] getIndexes() {
+        return this.indexes;
     }
 
     public List<Sign> getSigns() {
@@ -106,4 +515,39 @@ public abstract class AbstractSignManagement {
     public void setSigns(List<Sign> signs) {
         this.signs = signs;
     }
+
+    private enum ServiceInfoState {
+        STOPPED(0),
+        STARTING(1),
+        EMPTY_ONLINE(2),
+        ONLINE(3),
+        FULL_ONLINE(4);
+
+        private final int value;
+
+        ServiceInfoState(int value) {
+            this.value = value;
+        }
+
+        public int getValue() {
+            return this.value;
+        }
+    }
+
+    private static final class ServiceInfoSnapshotEntryComparator implements Comparator<Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>>> {
+
+        @Override
+        public int compare(Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>> o1, Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>> o2) {
+            return o1.getValue().getFirst().getServiceId().getName().compareTo(o2.getValue().getFirst().getServiceId().getName());
+        }
+    }
+
+    private static final class ServiceInfoSnapshotEntryComparator2 implements Comparator<Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>>> {
+
+        @Override
+        public int compare(Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>> o1, Map.Entry<UUID, Pair<ServiceInfoSnapshot, ServiceInfoState>> o2) {
+            return o1.getValue().getSecond().getValue() + o2.getValue().getSecond().getValue();
+        }
+    }
+
 }
