+++ cloudnet-modules/cloudnet-signs/src/main/java/de/dytanic/cloudnet/ext/signs/bukkit/BukkitCloudNetSignsPlugin.java
@@ -1,9 +1,9 @@
 package de.dytanic.cloudnet.ext.signs.bukkit;
 
 import de.dytanic.cloudnet.driver.CloudNetDriver;
+import de.dytanic.cloudnet.ext.signs.CloudNetSignListener;
 import de.dytanic.cloudnet.ext.signs.SignConfigurationEntry;
 import de.dytanic.cloudnet.ext.signs.bukkit.command.CommandCloudSign;
-import de.dytanic.cloudnet.ext.signs.bukkit.listener.BukkitCloudNetSignListener;
 import de.dytanic.cloudnet.ext.signs.bukkit.listener.BukkitSignInteractionListener;
 import de.dytanic.cloudnet.wrapper.Wrapper;
 import org.bukkit.Bukkit;
@@ -11,16 +11,6 @@ import org.bukkit.command.PluginCommand;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public final class BukkitCloudNetSignsPlugin extends JavaPlugin {
-
-    private static BukkitCloudNetSignsPlugin instance;
-
-    public BukkitCloudNetSignsPlugin() {
-        instance = this;
-    }
-
-    public static BukkitCloudNetSignsPlugin getInstance() {
-        return BukkitCloudNetSignsPlugin.instance;
-    }
 
     @Override
     public void onEnable() {
@@ -32,7 +22,7 @@ public final class BukkitCloudNetSignsPlugin extends JavaPlugin {
 
     @Override
     public void onDisable() {
-        CloudNetDriver.getInstance().getEventManager().unregisterListeners(getClassLoader());
+        CloudNetDriver.getInstance().getEventManager().unregisterListeners(super.getClassLoader());
         Wrapper.getInstance().unregisterPacketListenersByClassLoader(this.getClass().getClassLoader());
     }
 
@@ -48,7 +38,7 @@ public final class BukkitCloudNetSignsPlugin extends JavaPlugin {
         }
 
         //CloudNet listeners
-        CloudNetDriver.getInstance().getEventManager().registerListener(new BukkitCloudNetSignListener());
+        CloudNetDriver.getInstance().getEventManager().registerListener(new CloudNetSignListener());
 
         //Bukkit listeners
         Bukkit.getPluginManager().registerEvents(new BukkitSignInteractionListener(), this);
@@ -60,4 +50,5 @@ public final class BukkitCloudNetSignsPlugin extends JavaPlugin {
             Bukkit.getScheduler().scheduleSyncRepeatingTask(this, new SignKnockbackRunnable(signConfigurationEntry), 20, 5);
         }
     }
+
 }
